csci571/hw6/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
